chore(geometry): make page margins more similar to the LaTeX version (#63)
* chore(geometry): make page margins more similar to the LaTeX version

When I compare the compiled documents from this and the original
[egraff/uit-thesis](https://github.com/egraff/uit-thesis) repo, I find
that adjusting the page margins and the vertical space accordingly makes
them more similar. But this, of course, is only a matter of taste which
is better/more appropriate.

* fix(geometry): center the supervisors page text

The narrower general geometry is too narrow for the current supervisors
page. This commit fixes this by centring the text on that specific page.

Signed-off-by: engeir <38856990+engeir@users.noreply.github.com>

* fix: tweak supervisors layout

avoid "External supervisor:" breaking to a new line

---------

Signed-off-by: engeir <38856990+engeir@users.noreply.github.com>
Co-authored-by: Ole Tytlandsvik <ole.tytlandsvik@gmail.com>